Updated the map system to work with the new core library.
