+++ res/tiles/manifest.json
@@ -1,1 +1,1 @@
-{"files":{"battle.bin":{"inputPath":"assets/tiles/battle.png","tiles":80},"beholder.bin":{"inputPath":"assets/tiles/beholder.png","tiles":98},"dungeon.bin":{"inputPath":"assets/tiles/dungeon.png","tiles":256},"font.bin":{"inputPath":"assets/tiles/font.png","tiles":128},"hero.bin":{"inputPath":"assets/tiles/hero.png","tiles":128},"kobold.bin":{"inputPath":"assets/tiles/kobold.png","tiles":98},"objects.bin":{"inputPath":"assets/tiles/objects.png","tiles":128},"world_map.bin":{"inputPath":"assets/tiles/world_map.png","tiles":256}}}
+{"files":{"battle.bin":{"inputPath":"assets/tiles/battle.png","tiles":80},"beholder.bin":{"inputPath":"assets/tiles/beholder.png","tiles":98},"dungeon.bin":{"inputPath":"assets/tiles/dungeon.png","tiles":256},"font.bin":{"inputPath":"assets/tiles/font.png","tiles":128},"hero.bin":{"inputPath":"assets/tiles/hero.png","tiles":96},"kobold.bin":{"inputPath":"assets/tiles/kobold.png","tiles":98},"objects.bin":{"inputPath":"assets/tiles/objects.png","tiles":128},"world_map.bin":{"inputPath":"assets/tiles/world_map.png","tiles":256}}}
